Minor bug fixes
